A simple project I did to Learn Redux
This is a simple project I implemented to learn Redux on. It has a companion back-end project on Node.js, Express server and Mongodb.

It's not yet complete. When it's done, a user should be able to login create/delete projects and add/remove/edit Tasks within the project and later be able to add contributors and assign them tasks.